Sort supported video resolutions list before using it.
Originally, the code assume that the list is in ascending order. But as
some device (e.g. LG L90 Dual) send this list in descending order, and
actually we probably can't rely on it's order anyway, I decided to sort
the list before using it. 

This fixes 2 issues: 
1.) video resolution ordering in UI being inconsistent.
2.) when a device send this list in descending order, and unsupported
resolution is currently selected, it'll select lowest resolution,
which is sometimes actually too low for encoder. This cause recording
to fail.